requirements.txt
google-adk==1.21.0
google-genai==1.55.0
python-dotenv==1.2.1
uvicorn==0.38.0
fastapi==0.123.10